+++ export/CHANGELOG.md
@@ -12,7 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `orbax_export.typing` API
 - Require jax version >=0.4.30 as the jax.export API was introduced in this release.
 - Add version flag to export_manager.py and a stubbed class to handle each version.
-
+- Adds a new module base class and two new module subclasses one for TensorFlow and
+  one for Orbax.
 
 ## [0.0.4] - 2024-1-19